pyproject.toml
[tool.poetry]
name = "qna-bot"
version = "0.1.0"
description = ""
authors = ["RAHUL-COD3 <rahul@promactinfo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
azure-search-documents = "^11.5.1"
azure-identity = "^1.17.1"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
